fixed bug
When a user selected a repo to import, the index would be off because I picked
the index sequentially from the `store.discovered_repos` and that was being
shuffled around to disply, which screwed up the index. Now, we're indexing off
the username/repo combo, which works around the above.